+++ _private/js/llms-admin.js
@@ -52,11 +52,21 @@
 
 	$.fn.llmsStudentsSelect2 = function( options ) {
 
-		var defaults = {
-			multiple: false,
-			placeholder: 'Select a student',
-			width: '100%',
-		};
+		var self = this,
+			options = options || {},
+			defaults = {
+				enrolled_in: '',
+				multiple: false,
+				not_enrolled_in: '',
+				placeholder: 'Select a student',
+				width: '100%',
+			};
+
+		$.each( defaults, function( setting ) {
+			if ( self.attr( 'data-' + setting ) ) {
+				options[ setting ] = self.attr( 'data-' + setting );
+			}
+		} );
 
 		options = $.extend( defaults, options );
 
@@ -69,10 +79,12 @@
 				url: window.ajaxurl,
 				data: function( params ) {
 					return {
-						term: params.term,
-						page: params.page,
-						action: 'query_students',
 						_ajax_nonce: wp_ajax_data.nonce,
+						action: 'query_students',
+						page: params.page,
+						not_enrolled_in: params.not_enrolled_in || options.not_enrolled_in,
+						enrolled_in: params.enrolled_in || options.enrolled_in,
+						term: params.term,
 					};
 				},
 				processResults: function( data, params ) {